fixed services

+++ ShowerShow/Service/BlobStorageService.cs
@@ -9,19 +9,26 @@ namespace ShowerShow.Service
     public class BlobStorageService : IBlobStorageService
     {
         private IBlobStorageRepository blobStorageRepository;
+        private IUserService userService;
 
-        public BlobStorageService(IBlobStorageRepository blobStorageRepository)
+        public BlobStorageService(IBlobStorageRepository blobStorageRepository,IUserService userService)
         {
             this.blobStorageRepository = blobStorageRepository;
+            this.userService = userService;
         }
 
         public async Task DeleteProfilePicture(Guid userId)
         {
+            if (!await userService.CheckIfUserExistAndActive(userId))
+                throw new ArgumentException("The user does not exist or is inactive.");
             await blobStorageRepository.DeleteProfilePicture(userId);
         }
 
         public async Task<HttpResponseData> GetProfilePictureOfUser(HttpResponseData response, Guid userId)
         {
+            if (!await userService.CheckIfUserExistAndActive(userId))
+                throw new ArgumentException("The user does not exist or is inactive.");
+
             return await blobStorageRepository.GetProfilePictureOfUser(response, userId);
         }
 
@@ -32,6 +39,9 @@ namespace ShowerShow.Service
 
         public async Task UploadProfilePicture(Stream requestBody, Guid userId)
         {
+            if (!await userService.CheckIfUserExistAndActive(userId))
+                throw new ArgumentException("The user does not exist or is inactive.");
+
             await blobStorageRepository.UploadProfilePicture(requestBody, userId);
         }
