+++ ShowerShow/Service/ScheduleService.cs
@@ -10,10 +10,12 @@ namespace ShowerShow.Service
     public class ScheduleService : IScheduleService
     {
         private IScheduleRepository scheduleRepository;
+        private IUserService userService;
 
-        public ScheduleService(IScheduleRepository scheduleRepository)
+        public ScheduleService(IScheduleRepository scheduleRepository, IUserService userService)
         {
             this.scheduleRepository = scheduleRepository;
+            this.userService = userService;
         }
 
         public async Task AddScheduleToQueue(CreateScheduleDTO schedule, Guid userId)
@@ -23,11 +25,17 @@ namespace ShowerShow.Service
 
         public async Task CreateSchedule(Schedule schedule)
         {
+            if (!await userService.CheckIfUserExistAndActive(schedule.UserId))
+                throw new ArgumentException("The user does not exist or is inactive.");
+
             await scheduleRepository.CreateSchedule(schedule);
         }
 
         public async Task DeleteSchedule(Guid scheduleId)
         {
+            if (!await DoesScheduleExist(scheduleId))
+                throw new ArgumentException("The schedule does not exist.");
+
             await scheduleRepository.DeleteSchedule(scheduleId);
         }
 
@@ -38,16 +46,25 @@ namespace ShowerShow.Service
 
         public async Task<IEnumerable<Schedule>> GetAllSchedules(Guid UserId)
         {
+            if (!await userService.CheckIfUserExistAndActive(UserId))
+                throw new ArgumentException("The user does not exist or is inactive.");
+
             return await scheduleRepository.GetAllSchedules(UserId);
         }
 
         public async Task<Schedule> GetScheduleById(Guid scheduleId)
         {
+            if (!await DoesScheduleExist(scheduleId))
+                throw new ArgumentException("The schedule does not exist.");
+
             return await scheduleRepository.GetScheduleById(scheduleId);
         }
 
         public async Task<Schedule> UpdateSchedule(Guid scheduleId, UpdateScheduleDTO newSchedule)
         {
+            if (!await DoesScheduleExist(scheduleId))
+                throw new ArgumentException("The schedule does not exist.");
+
             return await scheduleRepository.UpdateSchedule(scheduleId, newSchedule);
         }
     }
